+++ MSGraph/functions/exchange/Get-MgaMailAttachment.ps1
@@ -59,6 +59,8 @@
         $Token
     )
     begin {
+        $requiredPermission = "Mail.Read"
+        $Token = Invoke-TokenScopeValidation -Token $Token -Scope $requiredPermission -FunctionName $MyInvocation.MyCommand
     }
 
     process {