+++ MSGraph/functions/exchange/Get-MgaMailMessage.ps1
@@ -104,6 +104,9 @@
         $Token
     )
     begin {
+        $requiredPermission = "Mail.Read"
+        $Token = Invoke-TokenScopeValidation -Token $Token -Scope $requiredPermission -FunctionName $MyInvocation.MyCommand
+
         $InvokeParams = @()
     }